requirements.txt
streamlit==1.37.1
langchain==0.2.12
openai==1.40.1
python-dotenv==1.0.1
faiss-cpu==1.8.0.post1
pandas==2.2.2
langchain-community

# python -m venv wisdom
# .\wisdom\Scripts\Activate.ps1
# Set-ExecutionPolicy -ExecutionPolicy RemoteSigned -Scope Process